+++ .github/CHANGELOG.md
@@ -11,7 +11,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/), 
  - Added an editor-only window to quickly retrieve an OAuth access token (menu : `Blizzard API/Get Access Token`)
    - This is mostly useful in Blizzard's documentation, by providing an access token to test endpoints
   
-## [0.45.0] - 2020-05-01
+## [0.46.0] - 2020-05-02
+### Added
+ - Added support for the following WoW Profile endpoint category :
+   - Character Statistics
+  
+## [0.45.0] - 2020-05-02
 ### Added
  - Added support for the following WoW Profile endpoint category :
    - Character Specializations